+++ src/app/(home)/page.tsx
@@ -1,8 +1,8 @@
 "use client"
 import { updateURLParams } from "@src/app/(home)/_utils/useDeterminePair"
 import { useDeterminePair } from "@src/app/(home)/_utils/useDeterminePair"
-import { SwapWidget } from "@src/components/DefuseSDK"
 import { getTokens } from "@src/components/DefuseSDK/features/machines/1cs"
+import { SwapWidget } from "@src/components/DefuseSDK/features/swap/components/SwapWidget"
 import { isBaseToken } from "@src/components/DefuseSDK/utils"
 import Paper from "@src/components/Paper"
 import { LIST_TOKENS } from "@src/constants/tokens"
