+++ src/app/otc/create-order/page.tsx
@@ -1,6 +1,6 @@
 "use client"
 import { useDeterminePair } from "@src/app/(home)/_utils/useDeterminePair"
-import { OtcMakerWidget } from "@src/components/DefuseSDK"
+import { OtcMakerWidget } from "@src/components/DefuseSDK/features/otcDesk/components/OtcMakerWidget"
 import Paper from "@src/components/Paper"
 import { LIST_TOKENS } from "@src/constants/tokens"
 import { useConnectWallet } from "@src/hooks/useConnectWallet"
